admin: consolidate Stories + Events into a Content surface (Phase 4)
- /admin/content: Upskiller Spotlights review/publish + the manual Luma
  events sync in one place. stories-admin.tsx and sync-events-button.tsx
  moved here.
- /admin/stories becomes a redirect stub to /admin/content.
- The admin landing (/admin) is now just the Cycles section: the ghost-button
  nav row (Invitations / All participants / Stories / Explore / Sync) is gone
  now that the sidebar owns navigation; CreateCycleForm stays. Retitled
  "Cycles".
- Nav Content item points at /admin/content.

Verified: tsc --noEmit, eslint, next build all green; zero changes under
moderator/chrome.
